fix: resolve all audit findings (33/33 tasks complete)
Security fixes:
- Fix broken rate limiter cleanup (TTL-based instead of AllowN(0))
- Prevent X-Forwarded-For IP spoofing (use RemoteAddr only)
- Require DB/Redis passwords via env vars (no defaults)
- Bind DB/Redis to localhost only (127.0.0.1)
- Change default SSLMode from disable to require
- Prevent directory listing on /output/* (images only)
- Add XSS sanitization for person names (html.EscapeString)
- Pin all GitHub Actions to SHA hashes

Stability fixes:
- Use context.Background() for pipeline (not r.Context())
- Fix data race in logger (atomic.Bool and atomic.Pointer)
- Add TTL cleanup for completed sessions (1 hour)
- Fix migration parser for PL/pgSQL functions ($ blocks)
- Fix TOCTOU race in detector anchor cache
- Fix graceful shutdown order (slice instead of map)
- Remove double signal handler registration

Performance improvements:
- Change Face.Embedding from float64 to float32 (matches pgvector)
- Use pgx.Batch for bulk face inserts (instead of N INSERTs)
- Optimize pixel processing (direct buffer access for NRGBA/RGBA)
- Add blob pooling in recognizer (sync.Pool)
- Remove duplicate IVFFlat index (keep HNSW only)

Architecture improvements:
- Remove duplicate code (pool.go, health.go, schema.sql ×2)
- Remove dead code (web.go, BBox.Array(), gracefulLogTimeout)
- Rename ML adapters: Repository → Gateway
- Fix scanner format mismatch (add webp, gif, bmp)
- Fix swallowed config validation error
- Add sqlc schema with HNSW index

Build:
- Fix non-existent Go version 1.25.6 → 1.24.0
- Run go mod tidy to clean go.sum

+++ internal/service/clustering/clustering_test.go
@@ -13,10 +13,10 @@ func TestClusterGroupsSimilarFaces(t *testing.T) {
 	t.Parallel()
 
 	faces := []model.Face{
-		{Embedding: []float64{1.0, 0.0, 0.0}},
-		{Embedding: []float64{0.99, 0.01, 0.0}},
-		{Embedding: []float64{-1.0, 0.0, 0.0}},
-		{Embedding: []float64{-0.98, -0.02, 0.0}},
+		{Embedding: []float32{1.0, 0.0, 0.0}},
+		{Embedding: []float32{0.99, 0.01, 0.0}},
+		{Embedding: []float32{-1.0, 0.0, 0.0}},
+		{Embedding: []float32{-0.98, -0.02, 0.0}},
 	}
 
 	clusters := Cluster(faces, 0.95)
@@ -39,9 +39,9 @@ func TestClusterAppliesTransitiveMerging(t *testing.T) {
 
 	// A~B and B~C above threshold, A~C below threshold.
 	faces := []model.Face{
-		{Embedding: []float64{1.0, 0.0}},
-		{Embedding: []float64{0.8, 0.6}},
-		{Embedding: []float64{0.28, 0.96}},
+		{Embedding: []float32{1.0, 0.0}},
+		{Embedding: []float32{0.8, 0.6}},
+		{Embedding: []float32{0.28, 0.96}},
 	}
 
 	clusters := Cluster(faces, 0.75)
@@ -82,14 +82,14 @@ func makeRandomFaces(n, dim int) []model.Face {
 	r := rand.New(rand.NewSource(42))
 	faces := make([]model.Face, n)
 	for i := 0; i < n; i++ {
-		emb := make([]float64, dim)
-		var norm float64
+		emb := make([]float32, dim)
+		var norm float32
 		for j := 0; j < dim; j++ {
-			v := r.Float64()*2 - 1
+			v := float32(r.Float64()*2 - 1)
 			emb[j] = v
 			norm += v * v
 		}
-		norm = math.Sqrt(norm)
+		norm = float32(math.Sqrt(float64(norm)))
 		if norm == 0 {
 			norm = 1
 		}
